Add files via upload
I'm starting the challenge for module 5 now for two reasons, 1. this challenge has a lot to do, and 2. I've already finished my daily planner. So, this is the next big thing. In this commit, I add some files that I've made. The HTML has the most in it. Right now, in this stage, everything is just a bunch of placeholders. but it's the frame that I'm trying to build. Then the JavaScript will go in and give it life. Like when God made Adam in the garden, He started with the mud sculpture, and then He breathed life into it making man.  I digress lol.