Update weather-script.js
In this commit, I got the API to display weather data on the page. I still need to add a search history and a five-day forecast. but at least it's displaying something.
Also, OpenWeatherMap API no longer supports the UV index, so I cannot display that value. The text box is working now if you type in a city and hit "Get the weather!", it will return the weather for that city.