Update weather-script.js
In this commit, I add the start to my search history.  I have it set up to create an array in local storage and add to that array every time I search for something. I'd like to say I have a plan on how I'm going to validate the onclick events but honestly, I have no clue how I'm going to do that.
Inch by inch life's a cinch.